refactor Move Adelaide time function to seperate file and combine shared lines
